project file strcuture change, small fixes
