+++ js/main.js
@@ -1,0 +1,9 @@
+import { renderCalendarCells } from './calendar.js';
+import { initModalEventListeners } from './modal.js';
+
+function init() {
+    renderCalendarCells();
+    initModalEventListeners();
+}
+
+init();
